// packages/client/src/types.ts
export type TileType =
  | 'go'
  | 'property'
  | 'railroad'
  | 'utility'
  | 'tax'
  | 'chance'
  | 'community_chest'
  | 'jail'
  | 'go_to_jail'
  | 'free_parking';

export type PropertyColor =
  | 'brown'
  | 'lightblue'
  | 'pink'
  | 'orange'
  | 'red'
  | 'yellow'
  | 'green'
  | 'darkblue';

export interface BaseTile {
  position: number;
  name: string;
  type: TileType;
}

export interface PropertyTile extends BaseTile {
  type: 'property';
  color: PropertyColor;
  price: number;
  housePrice: number;
  rent: [number, number, number, number, number, number];
  ownerId?: string;
  houses: number;
  mortgaged: boolean;
}

export interface RailroadTile extends BaseTile {
  type: 'railroad';
  price: number;
  ownerId?: string;
  mortgaged: boolean;
}

export interface UtilityTile extends BaseTile {
  type: 'utility';
  price: number;
  ownerId?: string;
  mortgaged: boolean;
}

export interface TaxTile extends BaseTile {
  type: 'tax';
  amount: number;
}

export type Tile =
  | BaseTile
  | PropertyTile
  | RailroadTile
  | UtilityTile
  | TaxTile;

export interface Player {
  id: string;
  name: string;
  color: string;
  money: number;
  position: number;
  inJail: boolean;
  jailTurns: number;
  isBankrupt: boolean;
  consecutiveDoubles: number;
}